+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 - Remove obsolete and unused run.php !3395 @\_fridtjof_
 - Provide CurrentUserRegionInterface for Session as independent implementation !3390 !3400 @KoeMai
 - Refactored settings page completely to vue and merged with foodsaver edit page for ambassador and orga user !3249 @chriswalg
+- Cleaned up the ApplicationGateway #1975 !3411 @alex.simm
 
 ## Dev/Test/CI stuff
 - Added backend support for getting, adding and updating achievements. #1724 !3317 !3318 @AntonBallmaier